requirements.txt
python-telegram-bot==20.0b0
requests
beautifulsoup4